qt6-build.eclass: drop IUSE=debug and QT6_{VERSION_MINOR,BUILD_DIR}
All these are currently unused by both the eclass and consumers.

* IUSE=debug:
This was used by qt5-build but is no-op here. There's a commented
out debug feature in qtwebengine:6, but if that is ever used it
should just add its own IUSE.

* QT6_VERSION_MINOR:
Undocumented and assume may be a unused left-over from writing
the case ${PV}. Dropping saves a $(ver_cut) fork in global scope.
Could add properly if ever felt needed, but for now...

* QT6_BUILD_DIR:
cmake.eclass' BUILD_DIR should be used instead, this variable is
 essentially just guessing what cmake will use and could be wrong
if an ebuild changes anything from cmake defaults.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>