qt6-build.eclass: support+unrestrict tests, export src_test+install
Qt5's may be a lost cause at this point wrt bug #457182, but can
do this fairly easily for Qt6 going forward (or at least for most
components, qtbase, qttools, and qtdeclarative are messier).

About src_install, it was defined but not exported. Not an issue
before but now we need it for cleanups.

Bug: https://bugs.gentoo.org/457182
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>