dev-qt/*: import qt6 live ebuilds from ::qt
As-is, no changes in this commit which is mostly the same
as the current in-tree ebuilds.

Maintaining these in two different repos feels like just a hassle
(to me), and would rather have everything in one place so I can
change live and release ebuilds simultaneously as needed, plus not
have to sync metadata or eclass changes either (plus chiitoo has
::gentoo commit access now).

May move packages if I happen to work on them, albeit I have no
intention to really touch Qt5 or LXQt (anyone working on these
are free to the same if they want, or keep current workflow).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>