dev-python/pyopencl: add 2023.1.2
Now with py3.12 support thanks to a patch which uses the canonical NumPy
way of getting the location of its header files instead of messing with
imp.find_module(). Upstream said in the relevant file they did it that
way because importing numpy "screws up distutils", however either this
is no longer the case with the current versions of everything involved
or PEP-517 installation is not affected.

Signed-off-by: Marek Szuba <marecki@gentoo.org>