dev-qt/qtwayland: skip failing test + two that need compositor (qt6)
seatv4 is perhaps due to newer libwayland, or the use of offscreen
rendering. But do not believe this would be used with any of the
mostly modern compositors we have packaged if I understand this
right, and so have not looked into this further (tst_seatv7 does pass).

Not fully looked into but the two others seem to be actually trying
to use a compositor (also opengl surfaces). We could run tinywl but
given the majority of other tests do not need that it does not seem
all that worth it (plus would likely be flaky).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>