app-office/libreoffice: bundle abseil-cpp
abseil isn't *really* intended to be a system library and having a year-old
pinned dep like this causes havoc for other packages and leads to uninstallable
combinations (e.g. libreoffice-bin + anything needing grpc).

Please see also the extensive discussion in the linked PRs wrt abseil now
depending on re2 and the ABI implications of the used C++ std. versions with
abseil. It's not worth it and there's more downsides to unbundling than not.

Bug: https://github.com/gentoo/gentoo/pull/31388
Bug: https://github.com/gentoo/gentoo/pull/31609
Bug: https://github.com/gentoo/gentoo/pull/31893
Bug: https://github.com/gentoo/gentoo/pull/32281
Bug: https://bugs.gentoo.org/908373
Closes: https://bugs.gentoo.org/912686
Signed-off-by: Sam James <sam@gentoo.org>