sys-power/acpid: phase out legacy LFS interfaces
acpid comes with its own implementation of isfdtype() for standard
libraries which do not define it, like musl. This implementation used
LFS interfaces, which are now deprecated and which since musl-1.2.4
no longer are a part of the _GNU_SOURCE feature set. musl documentation
says "just switch to standard interfaces" so let's do just that.

Closes: https://bugs.gentoo.org/906189
Signed-off-by: Marek Szuba <marecki@gentoo.org>