dev-qt/qtdeclarative: restrict tests (qt6)
Mostly had this figured out but the real issue is when
qtdeclarative is not installed, and this is a far worse
case than qttools or qtbase's few tests to skip.

QML imports are in general pretty messy to run from
tests as-is.

May revisit if the situation improves, albeit given believe
upstream installs before running tests it may not.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>