dev-qt/qttools: restrict tests (qt6)
Insists on running tools from the system (fails if qttools
is not installed), and even if force the paths (using qt.conf,
or seds), it still has problem finding support files and most
tests fail. Getting all this to work properly would likely
be too messy.

There's also an annoyance with FetchContent, albeit it'd be
easy to handle.

wrt clang, qdoc seems to produce correct output with >=clang-16
but the ordering is slightly altered (believed to be due to
.cpp parsing being bit different) and cause some tests to
fail simple diff comparison -- may not be worth worrying about.

Originally was skipping a lot of tests, but it was utltimately
not meaningful anymore (aka tools were not really tested).

May revisit if the situation improves, albeit given believe
upstream installs before running tests it may not.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>