dev-util/maturin: remove old maintainers
Pretty clear that I'm the one handling this right now,
may as well skip sending out unnecesary bug mail.

rust@ was added at the start but do not believe it was ever
meant to be there. Better alternative maintainer would likely
be python@ at this point if it becomes needed.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>