www-client/qutebrowser: remove widevine workaround for prefix
Now prexified in qtwebengine:6 itself, albeit the rare users
actually using this on prefix will need to manually rebuild
qtwebengine to get the change if <=qtwebengine-6.5.2.

Given this messed with tests did not want to keep this in
qutebrowser, and it can benefit others.

Technically don't need USE=widevine anymore, but prefer this
over a optfeature at the moment and qutebrowser is fast to
"rebuild" being python-only.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>