dev-qt/qtwebengine: do not install gn (qt6)
qtwebengine handling of gn does not allow to use CBUILD toolchain,
and only when cross-compiling it will look for system's. Also does
not respect *FLAGS as upstream intentially commented these out
because someone did a bug report with overzealous flags (QTBUG-64759).

And, unfortunately, qtwebengine's is modified and so dev-util/gn
may not work as expected with it (plus will need CMakeLists
modifications to be found).

Bare that, two possible options would be:
1. dev-qt/qtwebengine-gn or so to BDEPEND on and build cleanly
2. fix the mess that is the modified gen.py, and use BUILD_ variables

For now at least skip installing it as it'll just cause conflicts
if we want to do this later and is not really useful unless
want to do something like BDEPEND="cross? ( dev-qt/qtwebengine )".

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>