dev-qt/qt3d: optional qml+vulkan, tidy features/deps (qt6)
Most of these features are ON by default and add no deps, do not
believe much sense in keeping track of. And opengl+vulkan are
qtbase's features to handle and flipping can have unexpected
consequences.

system_zlib is only for bundled assimp and is default ON either
way (also taken from qtbase).

Technically we could make shadertools optional too but the RHI
renderer is the default+recommended in Qt6 and there is probably
little worth in exposing a USE.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>