dev-qt/qtdeclarative: match vulkan, move shader to bdep, adjust sql
Automagic happening with QT_CONFIG(vulkan), and libQt6Quick.so breaks
if vulkan is disabled on qtbase.

For shadertools this only use the "qsb" tool (thus why it checks for it
separately in CMakeLists), and such it would need to be BDEPEND. Odds
are a lot of fixing is needed to get cross right with qt6 still though
(i.e. some qtbase in BDEPENDs, maybe need QT_HOST_PATH).

Don't believe sql needs =, sql? should be enough.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>