dev-qt/qtmultimedia: tighten deps, make X+qml optional (qt6)
Hopefully right, kind of confusing.

Notes:
* qtbase[X=]: checks qtbase for wether to use xlib or not,
  we "could" probably force xlib off safely in "this" case
  but that still tend to be messy with Qt's ABI
* gst[X=]: similarly checks if enabled through defines, may
  possibly break without this if build against gst[X] then
  disable it
* qtbase[opengl=,vulkan=]: untested but source is literred in
  QT_CONFIG(opengl) and some (vulkan) and would most likely
  break if not matching like others (e.g. qtdeclarative) are
  known to
* ^ but for gstreamer, we can use gstreamer_gl feature
* qtshadertools is only used to call qsb that I can see, so
  move to BDEPEND, albeit does come with the caveat that
  (unlike qtdeclarative), it does not look for it separately
  to find it during cross-compilation
* qtsvg seems only used by examples, dropping may however have
  the downside that very few packages pull qtsvg and it is often
  needed by other applications (but it is wrong here)
* pulse[glib] dropped, may have been needed in the past but failing
  to see the need for USE=glib right now

X is enabled by default on qtbase, gst-plugins-base, and
here, so it should hopefully not cause mismatch headaches
even on non-desktop profiles (assuming users don't try to
manually enable X/opengl/vulkan per-packages). Pure wayland
systems should be able to avoid X11 entirely with Qt6.

Note that it "looks" for gstreamer even when it is disabled,
so it may give xorg-proto warnings (if missing) if gst was built
with USE=X. Non-issue given in that situation we don't use
gstreamer. If gst[-X], then xorg-proto is not needed at all.

Closes: https://bugs.gentoo.org/910364
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>