qt6-build.eclass: use upstream intended user_facing_tool_links.txt
Primary goal is to reduce the amount of things we need to track,
both what to install and when the symlink needs to be created
based on USE. Without this, a module could introduce a new tool
and may not even notice that it needs a symlink.

Essentially replaces qt6_symlink_binary_to_path which I do not
believe will have further uses (thus marking deprecated). That is,
unless we want to diverge from upstream intended links, but even
then may as well use a normal dosym in ebuilds.

Also re-arrange mycmakeargs a bit because yet again qttranslations
is the odd one out.

Bug: https://bugs.gentoo.org/863395
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>