games-emulation/pcsx2: update live, give up unbundling
We already had to go out of our way to unbundle a few extras, but
now upstream has removed all checks for alternatively using system
libraries (that someone had sent a PR for to add before) and this
is becoming too much of a pain to go against and very fragile.

May possibly revisit when it does a non-nightly release.

CMakeLists.txt set a very high lower bound on SDL2 for some reason,
but well, it's packaged.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>