app-emulation/wine-proton: warn if nvidia[abi_x86_32] is missing
Hard to know for sure if it's really used or not, but do not
want to introduce a kinda no-op VIDEO_CARDS on wine to actually
depend on it over a warning.

Less of an issue with mesa given other dependencies end up
requiring it (technically the dep is wrong given e.g. nvidia
would not need mesa[abi_x86_32], but well).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>