sys-kernel/gentoo-kernel-bin: update for gpkg with signed modules
and allow signing resulting kernel image for use with secureboot.

Stripping the modules is disabled, the modules in the gpkg are already stripped. Stripping after signing would remove the signature.

I suggest to introduce this change along with the next version bump so we don't have to deal with the situation where we have two different gpkgs for the same version.

See-also: projg2/binpkg-docker#2
See-also: #31988

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>