app-vim/jedi: add 0.11.2
Adds python3_11 support. In order to pass tests on python 3.11, the
minimum version of dev-python/jedi must be bumped to 0.19.

Due to mysterious reasons, the upstream version tag is now prefixed with
a "v", so the download url must adapt. While we are at it, naming the
distfile as "jedi" is simply wrong, since that could clash with
dev-python/jedi's distfiles if the version numbers ever collided. Use
MY_PN instead.

Closes: https://bugs.gentoo.org/882821
Closes: https://bugs.gentoo.org/896654
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>