sys-apps/dbus: add 1.15.8
Upstream have dropped non-meson support, so here we go!

Rebased the elogind patch reluctantly, but I've dropped the REQUIRED_USE
for one-of systemd/elogind, given floppym's analysis on the bug. The functionality
provided by elogind here is pretty minimal and it doesn't appear required at all,
so it doesn't make sense to force it.

Bug: https://bugs.gentoo.org/599494
Signed-off-by: Sam James <sam@gentoo.org>