dev-qt/qtspeech: new qt6 module, add 6.5.2 + live
Packaging some extra modules that are likely needed by upcoming
kde6 and were available with Qt5 in ::gentoo, so can be in a good
shape by the time need these for testing.

Please ask if need other modules that are still missing.

wrt IUSE=speechd, normally would do IUSE=speech, but this is more
specifically about enabling (optional) speechd support in a library
that can do tts by other means, and also do not want it to be confused
with enabling the "speech" Qt module either (nothing should need to
depend on this USE).

Note this module currently seems rather fragile (see ebuild comments),
considered just not packaging it (yet) but it will be needed and may
as well see where it goes now.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>