qt6-build.eclass: switch examples dir to /usr/lib*/qt6/examples
This is used to install compiled test executables (not source
examples or similar), meant to be executed from qtcreator
(not that I've tried it).

We do not install examples (yet), so no need to worry about files
or qtbase remembering a bad path.

Looking at other distros (at least those that do not set this to
a docdir like we were and there are many), putting this under
lib/*qt6/ alongside qt6/bin seems to be the typical choice.

For actual source code examples, Qt has opted to not support installing
these anymore (QTBUG-86302) so we'd have to doins -r ourselves either
way, likely like normal docs unless something expects to find these in
the Qt examples dir.

Bug: https://bugs.gentoo.org/881433
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>