dev-lang/python: fix 3.12 support for USE-disabled modules
python's build system doesn't provide straightforward support for
forcing modules to not be built. A couple different approaches have been
taken throughout the years, sometimes relying on gentoo-specific
patches, which are all broken in python 3.12 as it doesn't use distutils
and a setup.py anymore.

Adopt a new approach based on semi-documented internals of the CPython
build system.

Closes: https://bugs.gentoo.org/913596
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>