dev-qt/qtquick3d: depend on qtbase[opengl=,vulkan=]
Littered in QT_CONFIG(opengl/vulkan) and needed to avoid ABI issues.
Hopefully the last of these as it is annoying for users not handling
opengl/vulkan with global flags (and portage tend to give the wrong
suggestion here). Non-issue on desktop profiles by default.

[gles2-only=] is also a thing, but less of an issue given nothing
normally depends on that and simply left unset.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>