metadata/stabilization-groups: add qt/qt6.group
It may be extra work but planning to stabilize all qt6 modules
together not to repeat headaches users been having with qt5 and
mismatching keywords.

New modules being packaged should be added here as well unless
it is not possible to stabilize due to dependencies.

Note this is more or less a test file (or notes) currently.
Tooling may need time to understand disparities between slots
combined with ebuild quirks.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>