sys-block/thin-provisioning-tools: fix bug-911775
Many name clashes between boost::optional and std::optional are
resolved with explicit namespace qualification.  Because of a problem
with a ADL and a global "operator<<" function name not being included
in gtest's environment, they had to be defined in "namespace boost {".
Utterly hackish but only affects two testcase code files. Finally,
there was an issue of a "one-past the end" access causing a segfault
that was fixed by allocating an extra element to a zeroed std::vector.

Closes: https://bugs.gentoo.org/911775
Signed-off-by: Peter Levine <plevine457@gmail.com>