dev-qt/qtbase: optionally pdepend on qttranslations
For qt6, had bit of a draft to "merge" qttranslations with their
respective packages, plus a few splits for circular deps. But it
all started to feel churny and looked like it would be a pain to
maintain, prone to mistakes, and with very little to gain. Doing
our own tarballs could work but still requires some tracking and
writing scripts to handle it.

And then I do not think translations should need to be emerged
manually, it is something you can normally expect out of the box
unless you opt-out.

So let's just do the "nls? ( )" in qt6 even if runtime-only IUSE
are kind of terrible on a large package like qtbase.

Not unconditional given it does bring in linguist as extra, and
translations are still a decent ~13MB installed.

Live-only for now, no need to trigger rebuilds over this.

Bug: https://bugs.gentoo.org/810802
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>