dev-qt/qt-docs: add 6.5.2_p202307080352 (new package for SLOT=6)
Debated IUSE="doc" on every dev-qt/* (more typical), but this feels
so much simpler if you just want about all docs (possibly without
emerging the actual module even, or needing clang+qdoc to build).

Not to mention qtbase, qttools, qtdeclarative, and qtshadertools
would still need a separate package to handle circular deps
(or cross in case of qttools, plus it tend to not like to
run its own not-installed tools). Not that it is a big deal to
handle, but it is still things I do not particularly want around.

Note that this is complete qt-docs, includes documentation for
unpackaged modules as well.

Closes: https://bugs.gentoo.org/881435
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>