qt6-build.eclass: abort build rather than skip on deps issues
Does not cover everything, there are some more "custom" checks
that may either abort or skip. This is for the generic top
level one that is the same across all of dev-qt/*:6

Afaik none of these should be able to trigger right now, but
not impossible been overlooking if it just been skipped.

Skipping+succeeding on tinderboxes that use minimal dependencies
may especially lead to confusing issues without this.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>