app-text/a2ps: fix a2ps-lpr-wrapper
Quoting the bug:
"""
There is an error on a2ps-lpr-wrapper and the file was not printed

If lp is installed, a2ps-lpr-wrapper will use it otherwise it falls back to lpr.
However, the #{lp} variable defines the printer specifier as -P. lpr does not take -P, it takes -d.

The specifier needs to differ depending on which backend is being used; -d for lp or -P for lpr.
"""

See also the savannah bug (https://savannah.gnu.org/bugs/?64047).

Closes: https://bugs.gentoo.org/913751
Signed-off-by: Sam James <sam@gentoo.org>