dev-qt/qtbase: remove IUSE=systemd
This is to control where log messages will be sent by default, and
I think enabling this is just going to be confusing. Off-by-default
(even if found), and cmake gives a warning that we should do special
handling for it. Plus no parity for openrc given we were not enabling
syslog as alternative.

Have not looked much, but not seem most distros enable this.

(also adjust disparity in the opengl option sorting)

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>