dev-qt/qtscxml: tighten deps, make qml optional
Missed that qml could be reasonably optional here.

And cannot find where it would need the rest outside examples.
Gui could technically be optional but would need its own USE.
Do not think it is worth it here unless someone wants this.

Maybe missing it, but failed to find anything at all using
opengl, it is listed in the docs but...?

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>