dev-qt/qtwebengine: rename IUSE=pdf to pdfium (qt6)
This is in part a message to future revdeps that having their
own USE=pdf will not work too well given it is default on
desktop profile and we do not really want to pull webengine
by default unless really wanted. Plus gives a way to disable
globally for arches where qtwebengine is not keyworded.

At worst this will become innacurate if qtpdf changes its
backing library.

IUSE is still in live-only so should not really disrupt.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>