dev-qt/qt-creator: add 11.0.2
Not going to list every changes given this ebuild is from scratch.

There are a few ebuilds for this going around, but they about all
have (IMO) overcustomization, with a lot of USE to juggle for both
users and maintainers (some handled through sed'ing cmake files) --
all while plugins can still be disabled at runtime either way.

So this is now mostly limited to options that introduce an extra
dependency (e.g. clang). "Some" of the longest to build plugins
do have dependencies, so this still allows to cut down build time
a decent amount (not that this is a short build still).

Other small notes:
* Default IUSEs match what Qt Creator enables by default at runtime
 (albeit default USE=-tracing could be interesting for faster builds).
* Left gumbo unbundling out for now given ultimate goal is to unbundle
  litehtml instead (which itself uses gumbo), but currently it requires
  a large patch for 0.8 support (likewise to unbundle on qttools).

wrt bug #905166, delaying cleanup given restoring keywords is
non-trivial and users may need time to adjust for Qt5->Qt6, but
can be dropped anytime anyway if llvm:14 cleanup is happening
(do not consider this a blocker).

Bug: https://bugs.gentoo.org/905166
Closes: https://bugs.gentoo.org/835042
Closes: https://bugs.gentoo.org/905417
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>