dev-qt/qttools: prevent litehtml automagic, add note
Not packaged so not really an issue, but if it is found odds are it
will end in build failure given this supports litehtml-0.6 and is
broken with (current) litehtml-0.8.

Thought this package it, but would rather wait until that's sorted out
rather than carry rather large patches. Not to mention need to patch
qt-creator for this as well, qlitehtml which wraps around it is not
setup to be unbundled.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>