qt6-build.eclass: include patch version in subslot for >=6.5.3
This matches Qt5 which had been doing the same.

While working on qt-creator bump, noticed that it broke with
qtquick3d-6.5.9999, which turned out because of changes in
private APIs that will land in 6.5.3. In other words, it is
going to need a rebuild when that lands but current SLOT=0/6.5
do not allow this. This could be a one-off issue but let's
stay safer when private APIs are abused.

This also keeps the 9999 in live subslot, so switching
between live and releases will trigger rebuilds too.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>