dev-python/PyQt6: support QtPdf with IUSE=pdfium.
QtPdf is enabled in qtwebengine-6.5.3+.

Not using IUSE=pdf given qtwebengine tend to need special treatment,
and this makes it easier to mask the USE globally on arches where it
is not supported. Furthermore IUSE=pdf is default on some profiles and
would pull the heavy qtwebengine for users that may not want it.

Revbumping at same time as 6.5.3 given binding operator will cause
a rebuild either way.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>