www-client/qutebrowser: skip chromium version test
Checks if the chromium version matches known versions recorded in the
test. Upstream updates as needed but that means failure for any older
qutebrowser version whenever there's a new major qtwebengine (aka,
currently fails with 6.6.9999 which reports 112 rather than 108, but
otherwise qutebrowser works just fine with it and upstream has updated
this in qutebrowser-9999).

Not really useful for us, let's just skip it rather than update.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>