www-apps/tt-rss: Bump snapshot to 20230901, PHP 8.2, improve permissions
You are now instructed to always add the PHP user the ttrssd group, unless
everything is to run as the web server user. This was necessary before, but my
earlier wording was ambiguous, if not entirely wrong.

The config.php file permissions are now also locked down to secure the database
credentials. It was previously world-readable.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>