app-text/podofo: add 0.10.1
Upstream moved to github and did a lot of cleanup. I have taken the
opportunity to drop lots of complicated logic that I don't understand
the need for, but which doesn't seem to be relevant to 0.10.x anyway.
This is effectively a brand-new ebuild.

Of particular note, the library soname is now incremented in
CMakeLists.txt (with real ABI tracking!) and is NOT simply the appended
project version in full. The slot has been adjusted accordingly.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Joonas Niilola <juippis@gentoo.org>