app-containers/containers-storage: removal
This is a Go(lang) library/module which is used by packages like podman & buildah. Library is built statically and ships with the package itself (vendered).
So this package will never be in any other package's {B,R,}DEPEND. man pages shipped with this package are also irrelavant.

The reason why this was added in tree is still questionable. Anyways this package did nothing whole life and will RIP now 💐

Closes: https://bugs.gentoo.org/show_bug.cgi?id=914456
Signed-off-by: Rahil Bhimjiani <rahil3108@gmail.com>