app-emulation/vkd3d: add 1.9
Also do a dirty fix for implicit configure check, not going to
spend much time here given this is only packaged for old wine-7.

May eventually mask USE=vkd3d on it rather than keep updating this
given it is unlikely dx12-users are not using a more recent Wine
and it is questionable how compatible these bumps are.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>