app-text/podofo: skip tests which require USE flags to succeed
These do not indicate anything wrong with podofo -- the verbose errors
are just things like "cannot find an Arial font" or "png / jpeg is not
supported", and that's very innocent.

For the Arial test in particular, passing the test also requires
`SearchFont("LiberationSans")` so must be a build dependency when
enabling both test and fontconfig. Consequently, we drop ~hppa since
that font isn't keyworded on hppa; it will have to be rekeyworded later.

One could wish that the upstream testsuite automatically skipped those
tests, but it is what it is...

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>