dev-libs/libgpg-error: Use different autogen.sh fix for cross-prefix
hprefixify uses ${EPREFIX}, which may not be executable or may not even exist at
all in a cross-prefix situation. Make autogen.sh execute using the same Bash
interpreter as the package manager. Unfortunately, there doesn't seem to be a
non-Bash-specific way to do this, so we cannot address this upstream.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>