app-text/podofo: suppress possibly impactful compiler fp handling
The tests fail on specific combinations of compiler optimizations due to
https://github.com/podofo/podofo/issues/103

In particular, at -march=x86-64-v3 levels and up, plus -O2. Setting
-ffp-contract=off makes the tests pass again... this *could* just be a
test data issue, but it may also be genuine library mis-handling. To be
on the safe side, just compile podofo such that it definitely passes its
tests, while waiting for an upstream response.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>