virtual/podofo-build: drop 0.10.1-r1
This virtual package was only ever used by calibre, and only in order to
depend on the union of podofo itself plus boost, see bug 503802 for
details.

Since podofo 0.10.1 does not use boost at all, the virtual package is
obsoleted by this podofo update. The one package that depends on it,
requires the old podofo (and will be version bumped to depend
exclusively on the new podofo, *without* the virtual package).

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>