dev-qt/qttools: enable assistant by default in live
Provides QtHelp, and lacking that by default is likely to be
an annoyance given the amount of packages that previously
depended on dev-qt/qthelp:5. No urgency to trigger a rebuild,
so queue'ing it for next release.

Technically requires no extras assuming default IUSE on qtbase,
albeit may eventually pull litehtml if unbundled.

Unfortunately does mean that the optfeature will nag about
qt-docs when it is primarily intended for developers.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>